+++ src/main/java/org/rhausner/projectmanagement/projectmanagementservice/model/Project.java
@@ -18,7 +18,7 @@ public class Project {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @Column(nullable = false)
     private String name;
     private String description;
@@ -47,7 +47,7 @@ public class Project {
      * @param startDate   the planned start date
      * @param endDate     the planned end date
      */
-    public Project(Integer id, String name, String description, LocalDate startDate, LocalDate endDate) {
+    public Project(Long id, String name, String description, LocalDate startDate, LocalDate endDate) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -56,11 +56,11 @@ public class Project {
         this.projectStatus = ProjectStatus.PLANNED;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
